pyproject.toml
[tool.poetry]
name = "pymafia"
version = "0.0.0"  # This is the standard placeholder for poetry-dynamic-versioning
description = "A Python module and bridge for reflecting KoLmafia's Java environment."
license = "MIT"
authors = ["MrFizzyBubbs <MrFizzyBubbs@protonmail.com>"]
readme = "README.md"
repository = "https://github.com/MrFizzyBubbs/pymafia"

[tool.poetry.dependencies]
python = "^3.10"
pyjnius = "^1.4.2"
wrapt = "^1.15.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.15.1"
black = "^23.3.0"
isort = "^5.10.1"
docformatter = "^1.5.0"
mypy = "^0.971"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[tool.poetry-dynamic-versioning]
enable = true
format = "{base}"
style = "pep440"
strict = true

[tool.pytest.ini_options]
pythonpath = "src/"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
